Moved demo spawn logic to its own script

+++ SkiesOfSteel/Assets/Scripts/StartNetwork.cs
@@ -28,7 +28,7 @@ public class StartNetwork : MonoBehaviour
     {
         NetworkManager.Singleton.StartServer();
 
-        gameManager.SetNumOfPlayers((ushort) numOfPlayers);
+        gameManager.SetNumOfPlayers(numOfPlayers);
     }
 
     public void StartClient()
